requirements.txt
Django==3.1.2
djangorestframework==3.12.1
Markdown==3.3.3
django-filter==2.4.0
Pillow==8.0.1
# djangorestframework-simplejwt==4.6.0
knox==0.1.11
django-rest-knox==4.1.0
Jinja2==2.11.2